refactor: Application.java에 존재하는 game 관심사 BaseBallGame 클래스로 이동
