+++ src/main/java/baseball/domain/BaseBallGame.java
@@ -1,5 +1,9 @@
 package baseball.domain;
 
+import baseball.common.RandomNumGenerator;
+import baseball.view.EntryView;
+import baseball.view.ResultView;
+
 import java.util.List;
 
 public class BaseBallGame {
@@ -7,27 +11,48 @@ public class BaseBallGame {
     private List<Integer> inputNumbers;
     private int strikeCount = 0;
     private int ballCount = 0;
-    private boolean isEnded = false;
 
     public BaseBallGame(List<Integer> opponentNumbers, List<Integer> inputNumbers) {
         this.opponentNumbers = opponentNumbers;
         this.inputNumbers = inputNumbers;
     }
 
-    public void start() {
-        CompareNumbers();
+    public static void start() {
+        List<Integer> opponentNumbers = RandomNumGenerator.generate(1, 9, 3);
+        List<Integer> inputNumbers = EntryView.getInput();
+
+        BaseBallGame baseBallGame = new BaseBallGame(opponentNumbers, inputNumbers);
+        baseBallGame.play();
     }
 
-    public void retry(List<Integer> inputNumbers) {
-        this.inputNumbers = inputNumbers;
+    public void play() {
+        CompareNumbers();
+        ResultView.showResult(this);
+
+        if (strikeCount == 3) {
+            end();
+            return;
+        }
+
+        if (strikeCount < 3) {
+            retry();
+        }
+    }
+
+    public void retry() {
+        this.inputNumbers = EntryView.getInput();
         this.strikeCount = 0;
         this.ballCount = 0;
-        this.isEnded = false;
-        CompareNumbers();
+        play();
     }
 
     public void end() {
-        this.isEnded = true;
+        ResultView.showEnding(this);
+        String answer = ResultView.AskRestartOrTerminate();
+
+        if (answer.equals("1")) {
+            BaseBallGame.start();
+        }
     }
 
     private void CompareNumbers() {
@@ -61,13 +86,5 @@ public class BaseBallGame {
 
     public int getBallCount() {
         return ballCount;
-    }
-
-    public void setInputNumbers(List<Integer> inputNumbers) {
-        this.inputNumbers = inputNumbers;
-    }
-
-    public boolean isEnded() {
-        return this.isEnded;
     }
 }
